+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added width option to Panel
 - Change special method `__render_width__` to `__measure__`
 
+### Added
+
+- Added Console.show_cursor method
+
 ## [0.6.0] - 2020-03-03
 
 ### Added